Fixed issues with some stats
simplified the dailyAverage computation
refactored when I'll finish, and added copy for
the states where there is no data available

refs API#472
