+++ app/templates/components/stats-my-daily-average.hbs
@@ -1,5 +1,5 @@
 <div class="nw-card-header">
-  <div class="nw-card-title">My Daily {{goalUnitSingular}} Count</div>
+  <div class="nw-card-title">My Average {{goalUnitSingular}} Count</div>
 </div>
 <div class="nw-card-content">
   <div class="nw-card-icon"><img src="/images/cards/calendar.svg" /></div>
